requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
django-filter>=24.0
pymysql>=1.1.2
python-dotenv>=1.2.1
django-treebeard>=5.0.5
reportlab>=4.4.10
html5lib>=1.1
pypdf2>=3.0.1
pandas>=2.3.3
openpyxl>=3.1.5
Pillow>=10.4.0
celery>=5.6.2
redis>=7.2.1
requests>=2.32.5
python-dateutil>=2.9.0
whitenoise>=6.12.0
python-decouple>=3.8
dj-database-url>=2.3.0
